requirements.txt
apiflask==2.3.2
mysql-connector-python==9.2.0
flask-sqlalchemy==3.1.1
Flask-migrate==4.1.0
requests==2.31.0